+++ js/create-posts.js
@@ -14,7 +14,7 @@ import {
 import { createComment } from './create-comment.js';
 
 const createPostId = createRandomIdFromRangeGenerator(1, POST_COUNT);
-const createPostUrl = createRandomIdFromRangeGenerator(1, POST_COUNT);
+const createPostUrl = createRandomIdFromRangeGenerator(1, 25);
 const { POST_DESCRIPTION } = getMiniaturesData();
 
 const createPost = function () {
