+++ js/utils.js
@@ -6,12 +6,12 @@ function getRandomInteger (a, b) {
 }
 
 function createRandomIdFromRangeGenerator (min, max) {
-  const previousValues = [];
+  let previousValues = [];
 
   return function () {
     let currentValue = getRandomInteger(min, max);
     if (previousValues.length >= (max - min + 1)) {
-      return null;
+      previousValues = [];
     }
     while (previousValues.includes(currentValue)) {
       currentValue = getRandomInteger(min, max);
